# Pull command performance improvements