--- a/lean/commands/cloud/pull.py
+++ b/lean/commands/cloud/pull.py
@@ -31,16 +31,24 @@ def pull(project: Optional[str], pull_bootcamp: bool) -> None:
     """
     # Parse which projects need to be pulled
     project_id = None
+    project_name = None
     if project is not None:
         try:
             project_id = int(project)
         except ValueError:
-            pass
+            # We treat it as a name rather than an id
+            project_name = project
 
     api_client = container.api_client()
-    all_projects = api_client.projects.get_all()
-    project_manager = container.project_manager()
-    projects_to_pull = project_manager.get_projects_by_name_or_id(all_projects, project_id or project)
+    projects_to_pull = []
+    all_projects = None
+
+    if project_id is not None:
+        projects_to_pull.append(api_client.projects.get(project_id))
+    else:
+        all_projects = api_client.projects.get_all()
+        project_manager = container.project_manager()
+        projects_to_pull = project_manager.get_projects_by_name_or_id(all_projects, project_name)
 
     if project is None and not pull_bootcamp:
         projects_to_pull = [p for p in projects_to_pull if not p.name.startswith("Boot Camp/")]