Push command performance improvements by removing unnecessary API calls like `projects/read` (all) in some cases and `environments`

## Pull times before the improvements

### Newly created project

#### Pull 1:
```
lean-cli\root> lean cloud pull --project 13115718
[1/1] Pulling 'test_projects/Proyecto03'
Successfully pulled 'test_projects/Proyecto03/main.py'
Successfully pulled 'test_projects/Proyecto03/research.ipynb'
Finished pulling projects in 1.571234941482544 seconds
```

#### Pull 2:
```
lean-cli\root> lean cloud pull --project 13115718
[1/1] Pulling 'test_projects/Proyecto03'
Successfully pulled 'test_projects/Proyecto03/main.py'
Successfully pulled 'test_projects/Proyecto03/research.ipynb'
Finished pulling projects in 1.5493056774139404 seconds
```

#### Pull 3:
```
lean-cli\root> lean cloud pull --project 13115718
[1/1] Pulling 'test_projects/Proyecto03'
Successfully pulled 'test_projects/Proyecto03/main.py'
Successfully pulled 'test_projects/Proyecto03/research.ipynb'
Finished pulling projects in  1.5151171684265137 seconds
```

### Project with libraries

#### Pull 1:
```
lean-cli\root> lean cloud pull --project 12805396
[1/6] Pulling 'Library/PythonNewLib2'
[2/6] Pulling 'Library/TestLib'
[3/6] Pulling 'Library/TestLib2'
[4/6] Pulling 'Library/TestLib3'
[5/6] Pulling 'Library/TestLib4'
[6/6] Pulling 'test_projects/test_project'
Successfully pulled 'test_projects/test_project/main11.py'
Successfully pulled 'test_projects/test_project/main22.py'
Successfully pulled 'test_projects/test_project/main.py'
Successfully pulled 'test_projects/test_project/research.ipynb'
Adding/updating local library references to project TestLib2
[1/1] Adding/updating local library TestLib4 reference to project TestLib2
Adding/updating local library references to project test_project
[1/4] Adding/updating local library TestLib reference to project test_project
[2/4] Adding/updating local library TestLib2 reference to project test_project
[3/4] Adding/updating local library TestLib3 reference to project test_project
[4/4] Adding/updating local library PythonNewLib2 reference to project test_project
Finished pulling projects in 4.352985620498657 seconds
```

#### Pull 2:
```
lean-cli\root> lean cloud pull --project 12805396
[1/6] Pulling 'Library/PythonNewLib2'
[2/6] Pulling 'Library/TestLib'
[3/6] Pulling 'Library/TestLib2'
[4/6] Pulling 'Library/TestLib3'
[5/6] Pulling 'Library/TestLib4'
[6/6] Pulling 'test_projects/test_project'
Successfully pulled 'test_projects/test_project/main11.py'
Successfully pulled 'test_projects/test_project/main22.py'
Successfully pulled 'test_projects/test_project/main.py'
Successfully pulled 'test_projects/test_project/research.ipynb'
Adding/updating local library references to project TestLib2
[1/1] Adding/updating local library TestLib4 reference to project TestLib2
Adding/updating local library references to project test_project
[1/4] Adding/updating local library TestLib reference to project test_project
[2/4] Adding/updating local library TestLib2 reference to project test_project
[3/4] Adding/updating local library TestLib3 reference to project test_project
[4/4] Adding/updating local library PythonNewLib2 reference to project test_project
Finished pulling projects in 4.1878297328948975 seconds
```

#### Pull 3:
```
lean-cli\root> lean cloud pull --project 12805396
[1/6] Pulling 'Library/PythonNewLib2'
[2/6] Pulling 'Library/TestLib'
[3/6] Pulling 'Library/TestLib2'
[4/6] Pulling 'Library/TestLib3'
[5/6] Pulling 'Library/TestLib4'
[6/6] Pulling 'test_projects/test_project'
Successfully pulled 'test_projects/test_project/main11.py'
Successfully pulled 'test_projects/test_project/main22.py'
Successfully pulled 'test_projects/test_project/main.py'
Successfully pulled 'test_projects/test_project/research.ipynb'
Adding/updating local library references to project TestLib2
[1/1] Adding/updating local library TestLib4 reference to project TestLib2
Adding/updating local library references to project test_project
[1/4] Adding/updating local library TestLib reference to project test_project
[2/4] Adding/updating local library TestLib2 reference to project test_project
[3/4] Adding/updating local library TestLib3 reference to project test_project
[4/4] Adding/updating local library PythonNewLib2 reference to project test_project
Finished pulling projects in 4.6335289478302 seconds
```

## Pull times after the improvements

### Newly created project

#### Pull 1:
```
lean-cli\root> lean cloud pull --project 13115718
[1/1] Pulling 'test_projects/Proyecto03'
Successfully pulled 'test_projects/Proyecto03/main.py'
Successfully pulled 'test_projects/Proyecto03/research.ipynb'
Finished pulling projects in 1.1638734340667725 seconds
```

#### Pull 2:
```
lean-cli\root> lean cloud pull --project 13115718
[1/1] Pulling 'test_projects/Proyecto03'
Successfully pulled 'test_projects/Proyecto03/main.py'
Successfully pulled 'test_projects/Proyecto03/research.ipynb'
Finished pulling projects in 1.1376960277557373 seconds
```

#### Pull 3:
```
lean-cli\root> lean cloud pull --project 13115718
[1/1] Pulling 'test_projects/Proyecto03'
Successfully pulled 'test_projects/Proyecto03/main.py'
Successfully pulled 'test_projects/Proyecto03/research.ipynb'
Finished pulling projects in 1.0277948379516602 seconds
```

### Project with libraries

#### Pull 1:
```
lean-cli\root> lean cloud pull --project 12805396
[1/6] Pulling 'Library/PythonNewLib2'
[2/6] Pulling 'Library/TestLib'
[3/6] Pulling 'Library/TestLib2'
[4/6] Pulling 'Library/TestLib3'
[5/6] Pulling 'Library/TestLib4'
[6/6] Pulling 'test_projects/test_project'
Successfully pulled 'test_projects/test_project/main11.py'
Successfully pulled 'test_projects/test_project/main22.py'
Successfully pulled 'test_projects/test_project/main.py'
Successfully pulled 'test_projects/test_project/research.ipynb'
Adding/updating local library references to project TestLib2
[1/1] Adding/updating local library TestLib4 reference to project TestLib2
Adding/updating local library references to project test_project
[1/4] Adding/updating local library TestLib reference to project test_project
[2/4] Adding/updating local library TestLib2 reference to project test_project
[3/4] Adding/updating local library TestLib3 reference to project test_project
[4/4] Adding/updating local library PythonNewLib2 reference to project test_project
Finished pulling projects in 6.1534037590026855 seconds
```

#### Pull 2:
```
lean-cli\root> lean cloud pull --project 12805396
[1/6] Pulling 'Library/PythonNewLib2'
[2/6] Pulling 'Library/TestLib'
[3/6] Pulling 'Library/TestLib2'
[4/6] Pulling 'Library/TestLib3'
[5/6] Pulling 'Library/TestLib4'
[6/6] Pulling 'test_projects/test_project'
Successfully pulled 'test_projects/test_project/main11.py'
Successfully pulled 'test_projects/test_project/main22.py'
Successfully pulled 'test_projects/test_project/main.py'
Successfully pulled 'test_projects/test_project/research.ipynb'
Adding/updating local library references to project TestLib2
[1/1] Adding/updating local library TestLib4 reference to project TestLib2
Adding/updating local library references to project test_project
[1/4] Adding/updating local library TestLib reference to project test_project
[2/4] Adding/updating local library TestLib2 reference to project test_project
[3/4] Adding/updating local library TestLib3 reference to project test_project
[4/4] Adding/updating local library PythonNewLib2 reference to project test_project
Finished pulling projects in 5.9566733837127686 seconds
```

#### Pull 3:
```
lean-cli\root> lean cloud pull --project 12805396
[1/6] Pulling 'Library/PythonNewLib2'
[2/6] Pulling 'Library/TestLib'
[3/6] Pulling 'Library/TestLib2'
[4/6] Pulling 'Library/TestLib3'
[5/6] Pulling 'Library/TestLib4'
[6/6] Pulling 'test_projects/test_project'
Successfully pulled 'test_projects/test_project/main11.py'
Successfully pulled 'test_projects/test_project/main22.py'
Successfully pulled 'test_projects/test_project/main.py'
Successfully pulled 'test_projects/test_project/research.ipynb'
Adding/updating local library references to project TestLib2
[1/1] Adding/updating local library TestLib4 reference to project TestLib2
Adding/updating local library references to project test_project
[1/4] Adding/updating local library TestLib reference to project test_project
[2/4] Adding/updating local library TestLib2 reference to project test_project
[3/4] Adding/updating local library TestLib3 reference to project test_project
[4/4] Adding/updating local library PythonNewLib2 reference to project test_project
Finished pulling projects in 6.0950706005096436 seconds
```

**Note:** Times increased a bit for projects with several libraries because instead of doing one `projects.get_all()` we are doing N+1 `projects.get(project_id)` (1 for each library + 1 for the project). This is not bad because this is a special case: 90% of projects won't have libraries, 9% will have one and 1% will have more than one.